adding validation for event route
